git commit -m "Poczatek"

+++ src/Potwor.java
@@ -1,5 +1,5 @@
 public class Potwor extends Przeciwnik {
-    Potwor(int m,int l, double e, int hp,String t, boolean czypo){
+    Potwor(double m,int l, double e, double hp,String t, boolean czypo){
         moc = m;
         lvl = l;
         exp = e;
